Update vehicle.cpp
This modified version of the Gauge class uses a RoundGaugeGraphicsObject object to display the data from the CAN bus instead of using labels. 
The can_callback function updates the value of the round gauge by calling the setValue function, and the format_value and null_value functions are no longer needed.